+++ demo/src/main/java/com/orange/ods/demo/ui/components/cards/CardHorizontal.kt
@@ -23,15 +23,13 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import coil.compose.rememberAsyncImagePainter
-import coil.request.ImageRequest
-import coil.size.Size
-import com.orange.ods.compose.component.card.OdsTitleFirstCard
+import com.orange.ods.compose.component.card.OdsHorizontalCard
 import com.orange.ods.demo.R
 import com.orange.ods.demo.domain.recipes.LocalRecipes
 import com.orange.ods.demo.ui.components.utilities.clickOnElement
 
 @Composable
-fun CardTitleFirst(customizationState: CardCustomizationState) {
+fun CardHorizontal(customizationState: CardCustomizationState) {
     val context = LocalContext.current
     val recipes = LocalRecipes.current
     val recipe = rememberSaveable { recipes.filter { it.description.isNotBlank() }.random() }
@@ -46,19 +44,14 @@ fun CardTitleFirst(customizationState: CardCustomizationState) {
             val button1Text = stringResource(id = R.string.component_element_button1)
             val button2Text = stringResource(id = R.string.component_element_button2)
             val cardText = stringResource(id = R.string.component_card_element_card)
-            val imagePainter = rememberAsyncImagePainter(
-                model = ImageRequest.Builder(context)
-                    .data(recipe.imageUrl)
-                    .size(Size.ORIGINAL)
-                    .build(),
-                placeholder = painterResource(id = R.drawable.placeholder),
-                error = painterResource(id = R.drawable.placeholder)
-            )
 
-            OdsTitleFirstCard(
+            OdsHorizontalCard(
                 title = recipe.title,
-                image = imagePainter,
-                thumbnail = if (hasThumbnail) imagePainter else null,
+                image = rememberAsyncImagePainter(
+                    model = recipe.imageUrl,
+                    placeholder = painterResource(id = R.drawable.placeholder),
+                    error = painterResource(id = R.drawable.placeholder)
+                ),
                 subtitle = if (hasSubtitle) recipe.subtitle else null,
                 text = if (hasText) recipe.description else null,
                 onCardClick = if (isClickable) {
@@ -67,7 +60,9 @@ fun CardTitleFirst(customizationState: CardCustomizationState) {
                 button1Text = if (hasButton1) button1Text else null,
                 onButton1Click = { clickOnElement(context, button1Text) },
                 button2Text = if (hasButton2) button2Text else null,
-                onButton2Click = { clickOnElement(context, button2Text) }
+                onButton2Click = { clickOnElement(context, button2Text) },
+                imagePosition = imagePosition.value,
+                dividerEnabled = hasDivider
             )
         }
     }