+++ demo/src/main/java/com/orange/ods/demo/ui/components/cards/ComponentCard.kt
@@ -17,12 +17,16 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
+import com.orange.ods.compose.component.card.OdsHorizontalCardImagePosition
+import com.orange.ods.compose.component.chip.OdsChoiceChip
+import com.orange.ods.compose.component.chip.OdsChoiceChipsFlowRow
 import com.orange.ods.compose.component.list.OdsListItem
 import com.orange.ods.compose.component.list.OdsSwitchTrailing
 import com.orange.ods.demo.R
 import com.orange.ods.demo.ui.components.Variant
 import com.orange.ods.demo.ui.components.utilities.ComponentCountRow
 import com.orange.ods.demo.ui.components.utilities.ComponentCustomizationBottomSheetScaffold
+import com.orange.ods.demo.ui.utilities.composable.Subtitle
 
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
@@ -42,12 +46,22 @@ fun ComponentCard(variant: Variant) {
                         text = stringResource(id = R.string.component_element_thumbnail),
                         trailing = OdsSwitchTrailing(checked = thumbnailChecked)
                     )
+                } else if (variant == Variant.CardHorizontal) {
+                    Subtitle(textRes = R.string.component_card_horizontal_image_position, horizontalPadding = true)
+                    OdsChoiceChipsFlowRow(
+                        selectedChip = imagePosition,
+                        modifier = Modifier.padding(horizontal = dimensionResource(id = R.dimen.screen_horizontal_margin)),
+                        outlinedChips = true
+                    ) {
+                        OdsChoiceChip(textRes = R.string.component_card_horizontal_image_position_start, value = OdsHorizontalCardImagePosition.Start)
+                        OdsChoiceChip(textRes = R.string.component_card_horizontal_image_position_end, value = OdsHorizontalCardImagePosition.End)
+                    }
                 }
                 OdsListItem(
                     text = stringResource(id = R.string.component_element_subtitle),
                     trailing = OdsSwitchTrailing(checked = subtitleChecked)
                 )
-                if (variant in listOf(Variant.CardTitleFirst, Variant.CardImageFirst)) {
+                if (variant in listOf(Variant.CardTitleFirst, Variant.CardImageFirst, Variant.CardHorizontal)) {
                     OdsListItem(
                         text = stringResource(id = R.string.component_element_text),
                         trailing = OdsSwitchTrailing(checked = textChecked)
@@ -62,11 +76,19 @@ fun ComponentCard(variant: Variant) {
                         maxCount = CardCustomizationState.MaxActionButtonCount
                     )
                 }
+                if (variant == Variant.CardHorizontal) {
+                    if (!hasButton1) dividerChecked.value = false
+                    OdsListItem(
+                        text = stringResource(id = R.string.component_element_divider),
+                        trailing = OdsSwitchTrailing(checked = dividerChecked, enabled = hasButton1)
+                    )
+                }
             }) {
             when (variant) {
                 Variant.CardImageFirst -> CardImageFirst(customizationState = cardCustomizationState)
                 Variant.CardSmall -> CardSmall(customizationState = cardCustomizationState)
                 Variant.CardTitleFirst -> CardTitleFirst(customizationState = cardCustomizationState)
+                Variant.CardHorizontal -> CardHorizontal(customizationState = cardCustomizationState)
                 else -> {}
             }
         }